validate

+++ src/resolvers/index.ts
@@ -12,6 +12,7 @@ const resolvers = {
     // 登陆接口
     login: async (parent: any, param: any): Promise<String> => {
       try {
+        if(!param.userDetails.email) return "Email is required."
         let token: string = "";
         await User.findOne({ email: param.userDetails.email }).then(
           (res) => {
@@ -31,6 +32,8 @@ const resolvers = {
     //   注册接口 register + login return jwt token
     register: async (parent: any, param: any): Promise<String> => {
       try {
+        if(!param.userDetails.email) return "Email is required."
+        if(!param.userDetails.password) return "Password is required."
         let token: string = "";
         let password = await bcrypt.hash(param.userDetails.password, 10);
         await new User({
@@ -62,6 +65,7 @@ const resolvers = {
         return "";
       }
     },
+
     allPublicPosts: async (): Promise<IPost[]> => {
       try {
         let posts: IPost[] = [];
@@ -74,6 +78,7 @@ const resolvers = {
         return [];
       }
     },
+
     allPosts: async (parent: any, param: any): Promise<IPost[]> => {
       try {
         let posts: IPost[] = [];
@@ -87,37 +92,42 @@ const resolvers = {
         return [];
       }
     },
-    createPost: async (parent: any, param: any):  Promise<number> => {
+    createPost: async (parent: any, param: any):  Promise<String> => {
       try {
+        if(!param.post)return "Post cannot be empty"
+        if(!param.post.title)return "Title cannot be empty"
+        if(!param.post.body)return "Body cannot be empty"
         await new Post(param.post).save().then(res=>{
           console.log("new post created")
         })
-        return 1
+        return "Success"
       } catch (e) {
         console.log(e);
-        return 0
+        return "Failed"
       }
     },
-    updatePost: async (parent: any, param: any):  Promise<number> => {
+    updatePost: async (parent: any, param: any):  Promise<String> => {
       try {
+        if(!param._id) return "ID is required."
         Post.findByIdAndUpdate(param.post._id, param.post).then(res=>{
           console.log(" post updated")
         })
-        return 1
+        return "Success"
       } catch (e) {
         console.log(e);
-        return 0
+        return "Failed"
       }
     },
-    deletePost: async (parent: any, param: any):  Promise<number> => {
+    deletePost: async (parent: any, param: any):  Promise<String> => {
       try {
+        if(!param._id) return "ID is required."
         Post.findByIdAndDelete(param._id).then(res=>{
           console.log(" post deleted")
         })
-        return 1
+        return "Success"
       } catch (e) {
         console.log(e);
-        return 0
+        return "Failed"
       }
     },
   }
